requirements.dev.txt
flake8
mypy
django-stubs
djangorestframework-stubs
django-extensions
